backend/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-cors-headers==3.13.0
stackapi==0.2.0
django-redis==5.2.0
redis==4.3.4
django-ratelimit==3.0.1
redis-simple-cache